Split Calculator to seperate calss and use regex
This splits the calculator implementation to a seperate class.
Refactor implementation to use regular expressions to parser expression
Use Stream API when possible to incorperate a functional programming style.
Write test for the Calculator in calculatorTest class.
Fix bug associated with division with zero.
Add implementation to take care of +- and -- signs in the course of computation.